refactor: move all functionality out of benchmark_utils
