--- benchmark_utils/__init__.py
@@ -1,3 +1,0 @@
-from .utils import prox_fast_stack, prox_isotonic, preprocess_data
-
-__all__ = ["prox_fast_stack", "prox_isotonic", "preprocess_data"]
