+++ datasets/breheny.py
@@ -4,14 +4,14 @@ with safe_import_context() as import_ctx:
     import os
 
     import appdirs
-    import sklearn
     import numpy as np
     from download import download
     from rpy2 import robjects
     from rpy2.robjects import numpy2ri
     from scipy.sparse import csc_array
-
-    from benchmark_utils import preprocess_data
+    from scipy import sparse
+    from sklearn.feature_selection import VarianceThreshold
+    from sklearn.preprocessing import MaxAbsScaler, StandardScaler
 
 
 def fetch_breheny(dataset: str):
@@ -56,6 +56,13 @@ class Dataset(BaseDataset):
 
     def get_data(self):
         X, y = fetch_breheny(self.dataset)
-        X, y = preprocess_data(X, y, remove_zerovar=True, standardize=self.standardize)
+
+        X = VarianceThreshold().fit_transform(X)
+
+        if self.standardize:
+            if sparse.issparse(X):
+                X = MaxAbsScaler().fit_transform(X).tocsc()
+            else:
+                X = StandardScaler().fit_transform(X)
 
         return dict(X=X, y=y)
